requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
groq==0.9.0
httpx==0.25.2
PyPDF2==3.0.1
python-docx==0.8.11
scikit-learn==1.3.2
numpy==1.24.3
sqlalchemy==2.0.23
aiosqlite==0.19.0
python-dotenv==1.0.0
enum34==1.1.10